+++ apps/deploy-web/tests/fixture/base-test.ts
@@ -1,30 +1,59 @@
-import type { Page, Request, Route } from "@playwright/test";
+import type { Page } from "@playwright/test";
 import { test as baseTest } from "@playwright/test";
 
+import type { BrowserEnvConfig } from "@src/config/browser-env.config";
 import { testEnvConfig } from "./test-env.config";
 
 export * from "@playwright/test";
 
-export const test = baseTest.extend<{
-  page: Page;
-}>({
+export const test = baseTest.extend({
   page: async ({ page }, use) => {
-    await addUITestsToken(page);
+    await injectUIConfig(page);
     await use(page);
   }
 });
 
-export async function addUITestsToken(page: Page) {
-  const uiTestsToken = testEnvConfig.UI_TESTS_TOKEN;
-
-  if (uiTestsToken) {
-    await page.route(`${testEnvConfig.BASE_URL}/api/config`, async (route: Route, request: Request) => {
-      await route.continue({
-        headers: {
-          ...request.headers(),
-          "x-ui-tests-token": uiTestsToken
-        }
-      });
-    });
+export async function injectUIConfig(page: Page) {
+  if (!testEnvConfig.UI_CONFIG_SIGNATURE_PRIVATE_KEY) {
+    return;
   }
+
+  const uiConfig = await getSignedConfig(testEnvConfig.UI_CONFIG_SIGNATURE_PRIVATE_KEY);
+  await page.addInitScript(config => {
+    (window as any).__AK_INJECTED_CONFIG__ = config;
+  }, uiConfig);
+}
+
+const signedConfigCache = new Map<string, string>();
+async function getSignedConfig(privateKeyPem: string) {
+  if (signedConfigCache.has(privateKeyPem)) {
+    return signedConfigCache.get(privateKeyPem);
+  }
+
+  const config: Partial<BrowserEnvConfig> = {
+    // always pass token: https://deelopers.cloudflare.com/turnstile/troubleshooting/testing/#dummy-sitekeys-and-secret-keys
+    NEXT_PUBLIC_TURNSTILE_SITE_KEY: "1x00000000000000000000AA"
+  };
+  const serializedConfig = JSON.stringify(config);
+  const privateKey = await importPrivateKey(privateKeyPem);
+  const signatureBuffer = await crypto.subtle.sign("RSASSA-PKCS1-v1_5", privateKey, Buffer.from(serializedConfig));
+
+  const result = `${serializedConfig}.${Buffer.from(signatureBuffer).toString("base64")}`;
+  signedConfigCache.set(privateKeyPem, result);
+  return result;
+}
+
+async function importPrivateKey(pem: string) {
+  const der = Buffer.from(pem.replace(/-----[^-]+-----/g, "").replace(/\s+/g, ""), "base64");
+
+  return await crypto.subtle.importKey(
+    "pkcs8",
+    der,
+    {
+      name: "RSASSA-PKCS1-v1_5",
+      hash: "SHA-256"
+    },
+    false,
+    ["sign"]
+  );
 }